backend/foodgram/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.1
asn1crypto==1.5.1
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.2.1
django-cleanup==7.0.0
django-extensions==3.2.3
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-extra-fields==3.5.0
filetype==1.2.0
flake8==6.0.0
fonttools==4.40.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
lxml==4.9.2
mccabe==0.7.0
oauthlib==3.2.2
oscrypto==1.3.0
Pillow==9.5.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pydyf==0.6.0
pyflakes==3.0.1
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
PyJWT==2.7.0
pypdf==3.9.1
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.6.2
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.2
webencodings==0.5.1
xhtml2pdf==0.2.11
zopfli==0.2.2
